backend/requirements.txt
fastapi
uvicorn
qdrant-client
sqlalchemy
sqlmodel
psycopg[binary]
python-multipart
pydantic-settings
langchain
langchain-community
langchain-openai
openai
pypdf
python-docx
python-dotenv
aiofiles
torch
torchaudio
httpx==0.27.0
# New dependencies for local processing
openai-whisper
gTTS
